requirements.txt
unstructured[local-inference]
ebooklib
python-docx
pypandoc
python-dotenv   # if you need API keys in the future
